Sieve of Eratosthenes [Algorithm]
The Sieve of Eratosthenes is an ancient algorithm for finding all prime numbers up to a specified integer. The process involves iteratively marking as composite (i.e., not prime) the multiples of each prime number, starting with the first prime, 2. The numbers that remain unmarked at the end are the prime numbers within the given range.